apps/bench/server/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
pyyaml>=6.0
numpy<2
onnxruntime>=1.16
soundfile>=0.12
